split: extract AdminUserService from AdminService
Moves createUser, resendInvite, get/list/countUsers, toggle/setUserEnabled,
updateUser, revokeAllSessions, sendTestEmail to a focused service.
AdminService drops sessionRepository and emailPort dependencies.
