+++ src/main/kotlin/com/kauth/adapter/web/admin/AdminSessionAuditRoutes.kt
@@ -22,6 +22,7 @@ fun Route.adminSessionAuditRoutes(
     sessionRepository: SessionRepository,
     auditLogRepository: AuditLogRepository,
     adminService: AdminService,
+    adminUserService: com.kauth.domain.service.AdminUserService,
     applicationRepository: ApplicationRepository,
 ) {
     // -------------------------------------------------------------------
@@ -35,7 +36,7 @@ fun Route.adminSessionAuditRoutes(
         val totalCount = sessionRepository.countActiveByTenant(workspace.id)
         val sessions = sessionRepository.findActiveByTenant(workspace.id, limit = sessionsLimit)
         val sessionUserIds = sessions.mapNotNull { it.userId }.distinct()
-        val sessionUserMap = resolveUsernames(sessionUserIds, workspace.id, adminService)
+        val sessionUserMap = resolveUsernames(sessionUserIds, workspace.id, adminUserService)
         val sessionClientIds = sessions.mapNotNull { it.clientId }.distinct()
         val sessionClientMap = resolveClientNames(sessionClientIds, applicationRepository)
         val wsPairs = call.attributes[WsPairsAttr]
@@ -67,7 +68,7 @@ fun Route.adminSessionAuditRoutes(
 
     post("/sessions/revoke-all") {
         val workspace = call.attributes[WorkspaceAttr]
-        adminService.revokeAllSessions(workspace.id)
+        adminUserService.revokeAllSessions(workspace.id)
         call.respondRedirect("/admin/workspaces/${workspace.slug}/sessions?saved=revoked_all")
     }
 
@@ -95,7 +96,7 @@ fun Route.adminSessionAuditRoutes(
             )
         val total = auditLogRepository.countByTenant(workspace.id, eventType)
         val auditUserIds = events.mapNotNull { it.userId }.distinct()
-        val auditUserMap = resolveUsernames(auditUserIds, workspace.id, adminService)
+        val auditUserMap = resolveUsernames(auditUserIds, workspace.id, adminUserService)
         val auditClientIds = events.mapNotNull { it.clientId }.distinct()
         val auditClientMap = resolveClientNames(auditClientIds, applicationRepository)
         val auditClientLinks = resolveClientLinks(auditClientIds, applicationRepository)
