+++ src/main/kotlin/com/kauth/domain/service/AdminService.kt
@@ -7,15 +7,12 @@ import com.kauth.domain.model.AuditEventType
 import com.kauth.domain.model.RequiredAction
 import com.kauth.domain.model.Tenant
 import com.kauth.domain.model.TenantId
-import com.kauth.domain.model.User
 import com.kauth.domain.model.UserId
 import com.kauth.domain.port.ApplicationRepository
 import com.kauth.domain.port.AuditLogPort
 import com.kauth.domain.port.CorsPort
-import com.kauth.domain.port.EmailPort
 import com.kauth.domain.port.PasswordHasher
 import com.kauth.domain.port.PasswordPolicyPort
-import com.kauth.domain.port.SessionRepository
 import com.kauth.domain.port.TenantRepository
 import com.kauth.domain.port.UserRepository
 import com.kauth.domain.util.SecureTokens
@@ -26,10 +23,8 @@ class AdminService(
     private val applicationRepository: ApplicationRepository,
     private val passwordHasher: PasswordHasher,
     private val auditLog: AuditLogPort,
-    private val sessionRepository: SessionRepository,
     private val selfServiceService: UserSelfServiceService,
     private val passwordPolicy: PasswordPolicyPort? = null,
-    private val emailPort: EmailPort? = null,
     private val corsPort: CorsPort? = null,
 ) {
     private fun invalidateCors(tenantId: TenantId) {
@@ -59,326 +54,6 @@ class AdminService(
             null
         }
 
-    fun createUser(
-        tenantId: TenantId,
-        username: String,
-        email: String,
-        fullName: String,
-        password: String? = null,
-        sendInvite: Boolean = false,
-        baseUrl: String = "",
-    ): AdminResult<User> {
-        val tenant =
-            tenantRepository.findById(tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Workspace not found."))
-
-        if (username.isBlank()) {
-            return AdminResult.Failure(AdminError.Validation("Username is required."))
-        }
-        if (!username.matches(Regex("[a-zA-Z0-9._-]+"))) {
-            return AdminResult.Failure(
-                AdminError.Validation("Username may only contain letters, digits, dots, underscores, and hyphens."),
-            )
-        }
-        if (email.isBlank() || !email.contains('@')) {
-            return AdminResult.Failure(AdminError.Validation("A valid email address is required."))
-        }
-
-        // Resolve password hash and required actions based on invite mode
-        val resolvedPasswordHash: String
-        val resolvedRequiredActions: Set<RequiredAction>
-
-        if (sendInvite) {
-            resolvedPasswordHash = User.SENTINEL_PASSWORD_HASH
-            resolvedRequiredActions = setOf(RequiredAction.SET_PASSWORD)
-        } else {
-            if (!tenant.securityConfig.passwordLoginEnabled) {
-                return AdminResult.Failure(
-                    AdminError.Validation(
-                        "Password sign-in is disabled for this workspace. Send an invite instead.",
-                    ),
-                )
-            }
-            val pw =
-                password
-                    ?: return AdminResult.Failure(AdminError.Validation("Password is required."))
-            val policyError = passwordPolicy?.validate(pw, tenant)
-            if (policyError != null) {
-                return AdminResult.Failure(AdminError.Validation(policyError))
-            } else if (passwordPolicy == null && pw.length < tenant.passwordPolicyMinLength) {
-                return AdminResult.Failure(
-                    AdminError.Validation(
-                        "Password must be at least ${tenant.passwordPolicyMinLength} characters.",
-                    ),
-                )
-            }
-            resolvedPasswordHash = passwordHasher.hash(pw)
-            resolvedRequiredActions = emptySet()
-        }
-
-        if (userRepository.existsByUsername(tenantId, username)) {
-            return AdminResult.Failure(AdminError.Conflict("Username '$username' is already taken."))
-        }
-        if (userRepository.existsByEmail(tenantId, email)) {
-            return AdminResult.Failure(AdminError.Conflict("Email '${email.lowercase()}' is already registered."))
-        }
-
-        val user =
-            userRepository.save(
-                User(
-                    tenantId = tenantId,
-                    username = username.trim(),
-                    email = email.trim().lowercase(),
-                    fullName = fullName.trim(),
-                    passwordHash = resolvedPasswordHash,
-                    emailVerified = !sendInvite, // invite: verified on acceptance
-                    enabled = true,
-                    requiredActions = resolvedRequiredActions,
-                ),
-            )
-
-        // Record in password history (only for password-set users)
-        if (!sendInvite && passwordPolicy != null && tenant.passwordPolicyHistoryCount > 0) {
-            passwordPolicy.recordPasswordHistory(user.id!!, tenantId, resolvedPasswordHash)
-        }
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = user.id,
-                clientId = null,
-                eventType = AuditEventType.ADMIN_USER_CREATED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("username" to username, "invite" to sendInvite.toString()),
-            ),
-        )
-
-        // Send invite email — user is created regardless of email outcome.
-        // The email itself is async (fire-and-forget via emailScope).
-        // initiateInvite only fails synchronously if SMTP is not configured,
-        // which we already guard above with tenant.isSmtpReady.
-        if (sendInvite && tenant.isSmtpReady) {
-            when (selfServiceService.initiateInvite(user, tenant, baseUrl)) {
-                is SelfServiceResult.Success ->
-                    auditLog.record(
-                        AuditEvent(
-                            tenantId = tenantId,
-                            userId = user.id,
-                            clientId = null,
-                            eventType = AuditEventType.USER_INVITE_SENT,
-                            ipAddress = null,
-                            userAgent = null,
-                            details = mapOf("username" to username),
-                        ),
-                    )
-                is SelfServiceResult.Failure -> { /* SMTP guard above prevents this; log if it happens */ }
-            }
-        }
-
-        return AdminResult.Success(user)
-    }
-
-    fun resendInvite(
-        userId: UserId,
-        tenantId: TenantId,
-        baseUrl: String,
-    ): AdminResult<Unit> {
-        val tenant =
-            tenantRepository.findById(tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Workspace not found."))
-        val user =
-            userRepository.findById(userId, tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("User ${userId.value} not found."))
-
-        if (RequiredAction.SET_PASSWORD !in user.requiredActions) {
-            return AdminResult.Failure(AdminError.Validation("This user does not have a pending invite."))
-        }
-        if (!tenant.isSmtpReady) {
-            return AdminResult.Failure(AdminError.Validation("SMTP is not configured."))
-        }
-
-        return when (val result = selfServiceService.initiateInvite(user, tenant, baseUrl)) {
-            is SelfServiceResult.Success -> {
-                auditLog.record(
-                    AuditEvent(
-                        tenantId = tenantId,
-                        userId = userId,
-                        clientId = null,
-                        eventType = AuditEventType.USER_INVITE_SENT,
-                        ipAddress = null,
-                        userAgent = null,
-                        details = mapOf("username" to user.username, "action" to "resend"),
-                    ),
-                )
-                AdminResult.Success(Unit)
-            }
-            is SelfServiceResult.Failure ->
-                AdminResult.Failure(AdminError.Validation(result.error.message))
-        }
-    }
-
-    fun getUser(
-        userId: UserId,
-        tenantId: TenantId,
-    ): AdminResult<User> =
-        userRepository
-            .findById(userId, tenantId)
-            ?.let { AdminResult.Success(it) }
-            ?: AdminResult.Failure(AdminError.NotFound("User ${userId.value} not found."))
-
-    fun listUsers(
-        tenantId: TenantId,
-        search: String? = null,
-        limit: Int = Int.MAX_VALUE,
-        offset: Int = 0,
-    ): List<User> = userRepository.findByTenantId(tenantId, search, limit, offset)
-
-    fun countUsers(
-        tenantId: TenantId,
-        search: String? = null,
-    ): Long = userRepository.countByTenantId(tenantId, search)
-
-    fun toggleUserEnabled(
-        userId: UserId,
-        tenantId: TenantId,
-    ): AdminResult<Unit> {
-        val user =
-            userRepository.findById(userId, tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("User ${userId.value} not found."))
-        return setUserEnabled(userId, tenantId, !user.enabled)
-    }
-
-    fun sendTestEmail(
-        tenantId: TenantId,
-        recipientEmail: String,
-    ): AdminResult<Unit> {
-        val tenant =
-            tenantRepository.findById(tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Workspace not found."))
-        if (!tenant.isSmtpReady) {
-            return AdminResult.Failure(AdminError.Validation("SMTP is not configured for this workspace."))
-        }
-        val port =
-            emailPort
-                ?: return AdminResult.Failure(AdminError.Validation("Email delivery is not available."))
-        return try {
-            port.sendTestEmail(
-                to = recipientEmail,
-                workspaceName = tenant.displayName,
-                tenant = tenant,
-            )
-            auditLog.record(
-                AuditEvent(
-                    tenantId = tenantId,
-                    userId = null,
-                    clientId = null,
-                    eventType = AuditEventType.ADMIN_SMTP_TEST,
-                    ipAddress = null,
-                    userAgent = null,
-                    details = mapOf("recipient" to recipientEmail),
-                ),
-            )
-            AdminResult.Success(Unit)
-        } catch (e: Exception) {
-            AdminResult.Failure(AdminError.Validation("Failed to send test email: ${e.message}"))
-        }
-    }
-
-    fun revokeAllSessions(tenantId: TenantId): AdminResult<Int> {
-        val count = sessionRepository.revokeAllForTenant(tenantId)
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = null,
-                clientId = null,
-                eventType = AuditEventType.ADMIN_SESSIONS_REVOKED_ALL,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("sessionsRevoked" to count.toString()),
-            ),
-        )
-        return AdminResult.Success(count)
-    }
-
-    /**
-     * Updates a user's mutable profile fields (email, fullName).
-     * Username is intentionally immutable — it may appear in tokens already issued.
-     */
-    fun updateUser(
-        userId: UserId,
-        tenantId: TenantId,
-        email: String? = null,
-        fullName: String? = null,
-    ): AdminResult<User> {
-        val user =
-            userRepository.findById(userId, tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("User ${userId.value} not found."))
-
-        val resolvedEmail = email?.trim()?.lowercase() ?: user.email
-        val resolvedFullName = fullName?.trim() ?: user.fullName
-
-        if (resolvedEmail.isBlank() || !resolvedEmail.contains('@')) {
-            return AdminResult.Failure(AdminError.Validation("A valid email address is required."))
-        }
-
-        if (resolvedEmail != user.email && userRepository.existsByEmail(tenantId, resolvedEmail)) {
-            return AdminResult.Failure(AdminError.Conflict("Email '$resolvedEmail' is already registered."))
-        }
-
-        val updated = userRepository.update(user.copy(email = resolvedEmail, fullName = resolvedFullName))
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = userId,
-                clientId = null,
-                eventType = AuditEventType.ADMIN_USER_UPDATED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("username" to user.username),
-            ),
-        )
-
-        return AdminResult.Success(updated)
-    }
-
-    /**
-     * Enables or disables a user account.
-     */
-    fun setUserEnabled(
-        userId: UserId,
-        tenantId: TenantId,
-        enabled: Boolean,
-    ): AdminResult<Unit> {
-        val user =
-            userRepository.findById(userId, tenantId)
-                ?: return AdminResult.Failure(AdminError.NotFound("User ${userId.value} not found."))
-
-        userRepository.update(user.copy(enabled = enabled))
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = userId,
-                clientId = null,
-                eventType = if (enabled) AuditEventType.ADMIN_USER_ENABLED else AuditEventType.ADMIN_USER_DISABLED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("username" to user.username, "enabled" to enabled.toString()),
-            ),
-        )
-
-        return AdminResult.Success(Unit)
-    }
-
-    // =========================================================================
-    // Application management
-    // =========================================================================
-
-    /**
-     * Updates mutable application fields. clientId is immutable.
-     */
     fun updateApplication(
         appId: ApplicationId,
         tenantId: TenantId,
