+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminWebhooksTest.kt
@@ -131,7 +131,16 @@ class AdminWebhooksTest {
             applicationRepository = appRepo,
             passwordHasher = hasher,
             auditLog = auditLogPort,
+            selfServiceService = buildSelfService(),
+        )
+
+    private fun buildAdminUserService() =
+        com.kauth.domain.service.AdminUserService(
+            tenantRepository = tenantRepo,
+            userRepository = userRepo,
             sessionRepository = sessionRepo,
+            passwordHasher = hasher,
+            auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
         )
 
@@ -290,6 +299,7 @@ class AdminWebhooksTest {
                 workspaceSettingsService =
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
+                adminUserService = buildAdminUserService(),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
