+++ src/test/kotlin/com/kauth/adapter/web/api/ApiUserAttributeRoutesTest.kt
@@ -49,6 +49,19 @@ class ApiUserAttributeRoutesTest {
     private val userAttributeRepo = FakeUserAttributeRepository()
     private val hasher = FakePasswordHasher()
 
+    private fun buildFakeSelfService() =
+        com.kauth.domain.service.UserSelfServiceService(
+            userRepository = userRepo,
+            tenantRepository = tenantRepo,
+            sessionRepository = com.kauth.fakes.FakeSessionRepository(),
+            passwordHasher = hasher,
+            auditLog = com.kauth.fakes.FakeAuditLogPort(),
+            evTokenRepo = com.kauth.fakes.FakeEmailVerificationTokenRepository(),
+            prTokenRepo = com.kauth.fakes.FakePasswordResetTokenRepository(),
+            emailPort = com.kauth.fakes.FakeEmailPort(),
+            emailScope = kotlinx.coroutines.CoroutineScope(kotlinx.coroutines.Dispatchers.Unconfined),
+        )
+
     private val apiKeyService =
         ApiKeyService(apiKeyRepository = apiKeyRepo, tenantRepository = tenantRepo)
     private val userAttributeService =
@@ -352,20 +365,16 @@ class ApiUserAttributeRoutesTest {
                         applicationRepository = com.kauth.fakes.FakeApplicationRepository(),
                         passwordHasher = hasher,
                         auditLog = com.kauth.fakes.FakeAuditLogPort(),
+                        selfServiceService = buildFakeSelfService(),
+                    ),
+                adminUserService =
+                    com.kauth.domain.service.AdminUserService(
+                        tenantRepository = tenantRepo,
+                        userRepository = userRepo,
                         sessionRepository = com.kauth.fakes.FakeSessionRepository(),
-                        selfServiceService =
-                            com.kauth.domain.service.UserSelfServiceService(
-                                userRepository = userRepo,
-                                tenantRepository = tenantRepo,
-                                sessionRepository = com.kauth.fakes.FakeSessionRepository(),
-                                passwordHasher = hasher,
-                                auditLog = com.kauth.fakes.FakeAuditLogPort(),
-                                evTokenRepo = com.kauth.fakes.FakeEmailVerificationTokenRepository(),
-                                prTokenRepo = com.kauth.fakes.FakePasswordResetTokenRepository(),
-                                emailPort = com.kauth.fakes.FakeEmailPort(),
-                                emailScope =
-                                    kotlinx.coroutines.CoroutineScope(kotlinx.coroutines.Dispatchers.Unconfined),
-                            ),
+                        passwordHasher = hasher,
+                        auditLog = com.kauth.fakes.FakeAuditLogPort(),
+                        selfServiceService = buildFakeSelfService(),
                     ),
                 userAttributeService = userAttributeService,
                 claimMapperService = claimMapperService,
